Exception Handling

+++ src/main/java/com/desafio/crud_clientes/services/ClientService.java
@@ -3,10 +3,15 @@ package com.desafio.crud_clientes.services;
 import com.desafio.crud_clientes.dtos.ClientDTO;
 import com.desafio.crud_clientes.entities.Client;
 import com.desafio.crud_clientes.repositories.ClientRepository;
+import com.desafio.crud_clientes.services.exceptions.DatabaseException;
+import com.desafio.crud_clientes.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -20,7 +25,7 @@ public class ClientService {
 
     @Transactional(readOnly= true)
     public ClientDTO findById(Long id){
-       Client client = repository.findById(id).get();
+       Client client = repository.findById(id).orElseThrow(()-> new ResourceNotFoundException("Recurso não encontrado"));
        return new ClientDTO(client);
     }
 
@@ -40,15 +45,28 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update (Long id, ClientDTO dto){
-        Client entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ClientDTO(entity);
+        try {
+            Client entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ClientDTO(entity);
+        } catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+
+        }
     }
 
-    @Transactional
-    public void delete(Long id){
-        repository.deleteById(id);
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ClientDTO dto, Client entity){
